requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
automium==0.2.6
automium-web==0.1.1
bcrypt==4.1.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
fastapi-socketio==0.0.10
Flask==3.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
legacy==0.1.7
Mako==1.3.0
markdown2==2.4.10
MarkupSafe==2.1.3
nicegui==1.4.6
numpy==1.26.2
orjson==3.9.10
pandas==2.1.4
pscript==0.7.7
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-jose==3.3.0
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.14
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2023.3
uvicorn==0.22.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
